feat: refactored right sidebar to use shadcn ui
